수정:코드 각각 알맞는 컨트롤러로 옮김
