+++ src/main/java/com/kh/mng/bosspage/service/BossPageServiceImplSecond.java
@@ -75,65 +75,65 @@ public class BossPageServiceImplSecond implements BossPageServiceSecond{
 		return bossPageDaoSecond.deleteCouponKind(sqlSession, couponNo);
 	}
 
-	@Override
-	public int insertBoard(BoardEnroll board, BoardFileInfo boardfile) {
-		int result1 = bossPageDaoSecond.insertBoard(sqlSession, board);
-		int result2 = 1;
-		
-		//게시글 등록이 성공하고 첨부파일이 비어있지 않으면
-		if (result1 > 0 && boardfile.getChangeName() != null) {
-			result2 = bossPageDaoSecond.insertBoardAttachment(sqlSession, boardfile);
-		} 
-		
-//		else if (result1 > 0 && boardfile == null){
-//			result2 = 1;
-//		} else {
-//			result1 = result1 * 0;
+//	@Override
+//	public int insertBoard(BoardEnroll board, BoardFileInfo boardfile) {
+//		int result1 = bossPageDaoSecond.insertBoard(sqlSession, board);
+//		int result2 = 1;
+//		
+//		//게시글 등록이 성공하고 첨부파일이 비어있지 않으면
+//		if (result1 > 0 && boardfile.getChangeName() != null) {
+//			result2 = bossPageDaoSecond.insertBoardAttachment(sqlSession, boardfile);
+//		} 
+//		
+////		else if (result1 > 0 && boardfile == null){
+////			result2 = 1;
+////		} else {
+////			result1 = result1 * 0;
+////		}
+//		
+//		
+//		
+//		return result1 * result2;
+//	}
+
+	
+	
+	
+	
+//	// 핸드폰 중복 체크
+//	@Override
+//	public int checkPhoneNumber(String userPhone) {
+//		return bossPageDaoSecond.checkPhoneNumber(sqlSession, userPhone);
+//	}
+//	
+//	@Override
+//	@Transactional
+//	public int insertCertifyCode(PhoneSmsVo psv) {
+////		이 메소드에서만 sqlSession이 null로 뜨는 현상 발생 / 생성자 주입 방식을 통해 해결
+////		return bossPageDaoSecond.insertCertifyCode(sqlSession, psv);
+//		
+//		// 저장되어 있는 코드 있는지 전화번호로 확인
+//		int code = bossPageDaoSecond.selectCertifyCode(sqlSession, psv);
+//		int result1 = 1;
+//		int result2 = 1;
+//		
+//		// 저장된 코드 있는 경우 delete문 실행
+//		if (code > 0) {
+//			result1 = bossPageDaoSecond.deleteCertifyCode(sqlSession, psv);
 //		}
-		
-		
-		
-		return result1 * result2;
-	}
-
-	
-	
-	
-	
-	// 핸드폰 중복 체크
-	@Override
-	public int checkPhoneNumber(String userPhone) {
-		return bossPageDaoSecond.checkPhoneNumber(sqlSession, userPhone);
-	}
-	
-	@Override
-	@Transactional
-	public int insertCertifyCode(PhoneSmsVo psv) {
-//		이 메소드에서만 sqlSession이 null로 뜨는 현상 발생 / 생성자 주입 방식을 통해 해결
-//		return bossPageDaoSecond.insertCertifyCode(sqlSession, psv);
-		
-		// 저장되어 있는 코드 있는지 전화번호로 확인
-		int code = bossPageDaoSecond.selectCertifyCode(sqlSession, psv);
-		int result1 = 1;
-		int result2 = 1;
-		
-		// 저장된 코드 있는 경우 delete문 실행
-		if (code > 0) {
-			result1 = bossPageDaoSecond.deleteCertifyCode(sqlSession, psv);
-		}
-		
-		// 삭제 쿼리 실행 후 결과가 성공이면 insert문 실행
-		if (result1 > 0) {
-			result2 = bossPageDaoSecond.insertCertifyCode(sqlSession, psv);
-		}
-		
-		return result1 * result2;
-	}
-
-	@Override
-	public PhoneSmsVo checkCertifyCode(String phone) {
-		return bossPageDaoSecond.checkCertifyCode(sqlSession, phone);
-	}
+//		
+//		// 삭제 쿼리 실행 후 결과가 성공이면 insert문 실행
+//		if (result1 > 0) {
+//			result2 = bossPageDaoSecond.insertCertifyCode(sqlSession, psv);
+//		}
+//		
+//		return result1 * result2;
+//	}
+//
+//	@Override
+//	public PhoneSmsVo checkCertifyCode(String phone) {
+//		return bossPageDaoSecond.checkCertifyCode(sqlSession, phone);
+//	}
