+++ src/main/java/com/kh/mng/bosspage/service/BossPageServiceSecond.java
@@ -20,15 +20,15 @@ public interface BossPageServiceSecond {
 	// 쿠폰 삭제
 	int deleteCouponkind(int couponNo);
 	
-	// 게시글 등록
-	int insertBoard(BoardEnroll board, BoardFileInfo boardFile);
+//	// 게시글 등록
+//	int insertBoard(BoardEnroll board, BoardFileInfo boardFile);
 	
-	// 핸드폰 번호 중복 체크
-	int checkPhoneNumber(String userPhone);
-	
-	// 기존 인증코드 있는지 확인 및 삭제 후 암호화된 6자리 랜덤 숫자 저장
-	int insertCertifyCode(PhoneSmsVo psv);
-
-	// 저장된 핸드폰 번호로 된 인증번호 가져오기
-	PhoneSmsVo checkCertifyCode(String phone);
+//	// 핸드폰 번호 중복 체크
+//	int checkPhoneNumber(String userPhone);
+//	
+//	// 기존 인증코드 있는지 확인 및 삭제 후 암호화된 6자리 랜덤 숫자 저장
+//	int insertCertifyCode(PhoneSmsVo psv);
+//
+//	// 저장된 핸드폰 번호로 된 인증번호 가져오기
+//	PhoneSmsVo checkCertifyCode(String phone);
 }
