+++ src/main/java/com/kh/mng/community/controller/CommunityController.java
@@ -27,6 +27,8 @@ import com.google.gson.Gson;
 import com.kh.mng.common.model.vo.Attachment;
 import com.kh.mng.common.model.vo.PageInfo;
 import com.kh.mng.common.model.vo.Pagination;
+import com.kh.mng.community.model.dto.BoardEnroll;
+import com.kh.mng.community.model.dto.BoardFileInfo;
 import com.kh.mng.community.model.dto.BoardGoodInfo;
 import com.kh.mng.community.model.dto.BoardInfo;
 import com.kh.mng.community.model.dto.BoardPage;
@@ -393,6 +395,58 @@ public class CommunityController {
 	@GetMapping(value="enrollBoard.bo")
 	public String enrollBoard() {
 		return "community/writingPage";
+	}
+	
+//	커뮤니티 게시글 등록
+	
+	@PostMapping("enrollBoard.bo")
+	public String enrollBoard(BoardEnroll board, MultipartFile upfile, HttpSession session, Model model) {
+		BoardFileInfo boardFile = new BoardFileInfo();
+		Member loginUser = (Member)(session.getAttribute("loginUser"));
+
+		if(!upfile.getOriginalFilename().equals("")) {
+			String changeName = saveFile(upfile, session);
+			
+			boardFile.setOriginName(upfile.getOriginalFilename());
+			boardFile.setChangeName(changeName);
+			boardFile.setFilePath("resources/img/community/");
+			boardFile.setUserNo(loginUser.getUserNo());
+		}
+		
+		board.setUserNo(loginUser.getUserNo());
+		
+		int result = communityService.insertBoard(board, boardFile);
+		if(result > 0) { // 성공 -> list 페이지로 이동
+			session.setAttribute("alertMsg", "게시글 작성 성공");
+			return "redirect:/community";
+		} else { // 실패 -> 에러 페이지
+			model.addAttribute("errorMsg", "게시글 작성 실패");
+			return "";
+		}
+	}
+	
+	// 실제 넘어온 파일의 이름을 변경해서 서버에 저장하는 메소드
+	public String saveFile(MultipartFile upfile, HttpSession session) {
+		String originName = upfile.getOriginalFilename();
+		String currentTime = new SimpleDateFormat("yyyyMMddHHmmss").format(new Date());
+		int ranNum = (int)(Math.random() * 90000 ) + 10000;
+		String ext = originName.substring(originName.lastIndexOf("."));
+		
+		String changeName = currentTime + ranNum + ext;
+		
+		String savePath = session.getServletContext().getRealPath("resources/img/community/");
+	
+		try {
+			upfile.transferTo(new File(savePath + changeName));
+		} catch (IllegalStateException e) {
+			e.printStackTrace();
+		} catch (IOException e) {
+			e.printStackTrace();
+		}
+		
+//		경로 또는 url, 해당 파일 객체를 저장할 때 사용
+		
+		return changeName;
 	}
 	
 	//게시글 삭제 컨트롤러
