+++ src/main/java/com/kh/mng/community/service/CommunityService.java
@@ -7,6 +7,8 @@ import org.springframework.web.multipart.MultipartFile;
 
 import com.kh.mng.common.model.vo.Attachment;
 import com.kh.mng.common.model.vo.PageInfo;
+import com.kh.mng.community.model.dto.BoardEnroll;
+import com.kh.mng.community.model.dto.BoardFileInfo;
 import com.kh.mng.community.model.dto.BoardGoodInfo;
 import com.kh.mng.community.model.dto.BoardInfo;
 import com.kh.mng.community.model.dto.ReplyInfo;
@@ -69,5 +71,6 @@ public interface CommunityService {
 
 	int checkBoardOwner(int boardNo);
 
-
+	// 게시글 등록
+	int insertBoard(BoardEnroll board, BoardFileInfo boardFile);
 }
