+++ src/main/resources/mappers/community/community-board-mapper.xml
@@ -51,6 +51,27 @@
 	   		<result column="USER_NO"  property="userNo"></result>
 	   </resultMap>
 	   
+	   <resultMap type="BoardEnroll" id="boardEnrollResult">
+         <result column="BOARD_NO" property="boardNo"/>
+         <result column="BOARD_TITLE" property="boardTitle"/>
+         <result column="BOARD_CONTENT" property="boardContent"/>
+         <result column="COUNT" property="count"/>
+         <result column="CREATE_DATE" property="createDate"/>
+         <result column="EDIT_DATE" property="editDate"/>
+         <result column="STATUS" property="status"/>
+         <result column="USER_NO" property="userNo"/>
+         <result column="BOARD_CATEGORY_NO" property="boardCategoryNo"/>
+   	</resultMap>
+   	
+   	<resultMap type="BoardFileInfo" id="boardFileInfoResult">
+         <result column="PIC_NO" property="picNo"/>
+         <result column="ORIGIN_NAME" property="originName"/>
+         <result column="CHANGE_NAME" property="changeName"/>
+         <result column="FILE_PATH" property="filePath"/>
+         <result column="FILE_LEVEL" property="fileLevel"/>
+         <result column="USER_NO" property="userNo"/>
+         <result column="BOARD_NO" property="boardNo"/>
+   	</resultMap>
   
   	<select id="selectBoardList" resultMap="selectBoards">
   	  	SELECT B.BOARD_NO,
@@ -323,5 +344,46 @@
 	<select id="checkBoardOwner" resultType="_int">
 		SELECT USER_NO FROM BOARD WHERE BOARD_NO = #{boardNo}
 	</select>
+	
+	
+	
+	
+	
+	
+	
+	<insert id="insertBoard">
+   		INSERT INTO BOARD(
+   						BOARD_NO,
+   						BOARD_TITLE,
+   						BOARD_CONTENT,
+   						USER_NO,
+   						BOARD_CATEGORY_NO
+   						)
+   					VALUES(
+   						nextval('SEQ_BORD'),
+   						#{boardTitle},
+   						#{boardContent},
+   						#{userNo},
+   						#{boardCategoryNo}
+   						)
+   	</insert>
+   	
+   	<insert id="insertBoardAttachment">
+   		INSERT INTO PICTURE(
+   							PIC_NO, 
+   							ORIGIN_NAME, 
+   							CHANGE_NAME, 
+   							FILE_PATH,
+							BOARD_NO
+							)
+					 VALUES(
+							nextval('SEQ_PIC'),
+							#{originName},
+							#{changeName},
+							#{filePath},
+							currval('SEQ_BORD')
+							)
+							
+   	</insert>
    
   </mapper>
